ai-service/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
pypdf
pdfplumber
python-multipart
numpy
python-dotenv
torch